Tests | Test Improvements + Override tags in SQLServerDriver (#979)
