+++ src/test/java/com/microsoft/sqlserver/jdbc/SSLCertificateValidation.java
@@ -11,12 +11,14 @@ import static org.junit.Assert.assertTrue;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
 
 
 @RunWith(JUnitPlatform.class)
+@Tag("AzureDWTest")
 public class SSLCertificateValidation {
 
     /**
