+++ src/test/java/com/microsoft/sqlserver/jdbc/TestResource.java
@@ -47,6 +47,7 @@ public final class TestResource extends ListResourceBundle {
             {"R_createDropAlterTableFailed", "Create/drop/alter table with preparedStatement failed!"},
             {"R_grantFailed", "grant table with preparedStatement failed!"},
             {"R_connectionIsClosed", "The connection is closed."},
+            {"R_ConnectionURLNull", "The connection URL is null." },
             {"R_connectionIsNotClosed", "The connection is not closed."},
             {"R_invalidExceptionMessage", "Invalid exception message"},
             {"R_failedValidate", "failed to validate values in $0} "}, {"R_tableNotDropped", "table not dropped. "},
