+++ src/test/java/com/microsoft/sqlserver/jdbc/UtilTest.java
@@ -9,6 +9,7 @@ import static org.junit.Assert.assertEquals;
 import java.sql.SQLException;
 import java.util.UUID;
 
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
@@ -19,6 +20,7 @@ import org.junit.runner.RunWith;
  *
  */
 @RunWith(JUnitPlatform.class)
+@Tag("AzureDWTest")
 public class UtilTest {
 
     @Test
